Class Inventory
This class provides an ability to resolve name of datacenter
by hostname. There are two sources of this information are used:
a cocaine worker with handle 'get_dc_by_host', and cache saved in
MongoDB. Cache from DB is reloaded with a period specified in option
'infrastructure_dc_cache_update_period'. Default value is 150 sec.
Once in 'infrastructure_dc_cache_valid_time' (default is 7 days)
cocaine's handle is invoked and the database is updated in accordance.